email_auto_responder_flow/pyproject.toml
[project]
name = "email_auto_responder_flow"
version = "0.1.0"
description = "email_auto_responder_flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.11"
dependencies = [
    "crewai[tools]>=0.100.1,<1.0.0",
    "google-api-python-client>=2.86.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.scripts]
kickoff = "email_auto_responder_flow.main:kickoff"
plot = "email_auto_responder_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
